requirements.txt
apache-airflow
psycopg2-binary
celery
pyspark